backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
pydantic==2.5.2
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0
itsdangerous==2.1.2
polyline==1.4.0

# Google Cloud Services
google-cloud-aiplatform==1.38.1
# vertexai is included in google-cloud-aiplatform
google-cloud-language==2.13.0
googlemaps==4.10.0
google-cloud-texttospeech==2.14.1
google-cloud-storage==2.14.0
google-cloud-speech==2.20.0
google-cloud-secret-manager==2.16.2
google-cloud-logging==3.5.0
google-cloud-translate==3.11.2

# AI and ML
langchain==0.1.0
langchain-google-vertexai==0.0.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
joblib==1.3.2

# Spotify integration
spotipy==2.23.0

# Caching and performance
redis==5.0.1
aioredis==2.0.1
slowapi==0.1.9
# python-limiter removed - rate limiting handled by slowapi

# Third-party integrations
stripe==7.5.0
twilio==8.10.1
sendgrid==6.11.0

# Async support
aiofiles==23.2.1
aiohttp==3.9.1

# Security
cryptography==41.0.7
# python-owasp removed - not a valid package
pyotp==2.9.0
qrcode==7.4.2
bcrypt==4.1.2

# Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Database migrations
alembic==1.13.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
ruff==0.1.6
types-requests>=2.31.0.20231231